+++ Sources/AnyLintCLI/Tasks/LintTask.swift
@@ -4,6 +4,9 @@ import Utility
 
 struct LintTask {
     let configFilePath: String
+    let logDebugLevel: Bool
+    let failOnWarnings: Bool
+    let validateOnly: Bool
 }
 
 extension LintTask: TaskHandler {
@@ -25,8 +28,17 @@ extension LintTask: TaskHandler {
             log.message("Start linting using config file at \(configFilePath) ...", level: .info)
 
             var command = "\(configFilePath.absolutePath) \(log.outputType.rawValue)"
-            if log.logDebugLevel {
+
+            if logDebugLevel {
                 command += " \(Constants.debugArgument)"
+            }
+
+            if failOnWarnings {
+                command += " \(Constants.strictArgument)"
+            }
+
+            if validateOnly {
+                command += " \(Constants.validateArgument)"
             }
 
             try Task.run(bash: command)