+++ CHANGELOG.md
@@ -31,6 +31,13 @@ If needed, pluralize to `Tasks`, `PRs` or `Authors` and list multiple entries se
 ### Security
 - None.
 
+## [0.5.0] - 2020-04-22
+### Added
+- New `-s` / `--strict` option to fail on warnings as well (by default fails only on errors).  
+  PR: [#15](https://github.com/Flinesoft/AnyLint/pull/15) | Author: [Cihat Gündüz](https://github.com/Jeehut)
+- New `-l` / `--validate` option to only runs validations for `matchingExamples`, `nonMatchingExamples` and `autoCorrectExamples`.  
+  PR: [#17](https://github.com/Flinesoft/AnyLint/pull/17) | Author: [Cihat Gündüz](https://github.com/Jeehut)
+
 ## [0.4.0] - 2020-04-20
 ### Added
 - New `-d` / `--debug` option to log more info about what AnyLint is doing. Required to add a checks completion block in `logSummaryAndExit` and moved it up in the blank template.